# application/server/openapi.yaml
openapi: 3.0.3
info:
  title: Web Speed Hackathon 2026 API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
  /signup:
    post:
      tags:
        - Auth
      summary: サインアップ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - name
              properties:
                username:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - USERNAME_TAKEN
                      - INVALID_USERNAME
  /signin:
    post:
      tags:
        - Auth
      summary: サインイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
...
  /me:
    get:
      tags:
        - Users
      summary: ログインユーザー情報
      operationId: getCurrentUser
      security:
        - session: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Users
      summary: ログインユーザー情報更新
      operationId: updateCurrentUser
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{username}:
    get:
      tags:
        - Users
      summary: ユーザー取得
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{username}/posts:
    get:
      tags:
        - Users
      summary: ユーザーの投稿一覧
      operationId: getUserPosts
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: connect.sid
      description: express-session によるセッション Cookie
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
      description: 取得件数
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
      description: オフセット
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
    ProfileImage:
      type: object
      required: [id, alt]
      properties:
        id:
          type: string
          format: uuid
        alt:
          type: string
    User:
      type: object
      required: [id, username, name, description, createdAt, profileImage]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        profileImage:
          $ref: "#/components/schemas/ProfileImage"
    Image:
      type: object
      required: [id, alt]
      properties:
        id:
          type: string
          format: uuid
        alt:
          type: string
    Movie:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
    Sound:
      type: object
      required: [id, title, artist]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
    Post:
      type: object
      required: [id, text, createdAt, user, images]
      properties:
        id:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        movie:
          $ref: "#/components/schemas/Movie"
          nullable: true
        sound:
          $ref: "#/components/schemas/Sound"
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
...